fix OBJMaterlistfmt() and OBJfmt()
%g was being used to print the color values and vertex
parameters, but it's too accurate (can print wild values).
it seems standard with other exporters to print numbers up
to five or six decimal places, so we do five for materials
and six for vertex data.

a little bug was printing the wrong face format, getting
the different indices in the wrong order by not separating
them properly with a slash.